requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
google-cloud-firestore
google-auth
